+++ cppad/local/sub_op.hpp
@@ -164,8 +164,8 @@ inline void reverse_subvv_op(
 	CPPAD_ASSERT_UNKNOWN( d < nc_partial );
 
 	// Partial derivatives corresponding to arguments and result
-	Base* px = partial + arg[0] * nc_partial;
-	Base* py = partial + arg[1] * nc_partial;
+	Base* px = partial + size_t(arg[0]) * nc_partial;
+	Base* py = partial + size_t(arg[1]) * nc_partial;
 	Base* pz = partial + i_z    * nc_partial;
 
 	// number of indices to access
@@ -325,7 +325,7 @@ inline void reverse_subpv_op(
 	CPPAD_ASSERT_UNKNOWN( d < nc_partial );
 
 	// Partial derivatives corresponding to arguments and result
-	Base* py = partial + arg[1] * nc_partial;
+	Base* py = partial + size_t(arg[1]) * nc_partial;
 	Base* pz = partial + i_z    * nc_partial;
 
 	// number of indices to access
@@ -485,7 +485,7 @@ inline void reverse_subvp_op(
 	CPPAD_ASSERT_UNKNOWN( d < nc_partial );
 
 	// Partial derivatives corresponding to arguments and result
-	Base* px = partial + arg[0] * nc_partial;
+	Base* px = partial + size_t(arg[0]) * nc_partial;
 	Base* pz = partial + i_z    * nc_partial;
 
 	// number of indices to access
